requirements.txt
# Core API
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9

# LLM + embeddings
openai==1.50.2
tiktoken==0.7.0

# Retrieval
faiss-cpu==1.8.0.post1

# Parsing
pypdf==5.0.1
beautifulsoup4==4.12.3
python-docx==1.1.2
lxml==5.3.0

# Utilities
python-dotenv==1.0.1
numpy==2.1.0

# Evals & tests
pyyaml==6.0.2
pytest==8.3.2
httpx==0.27.2
